wip: changing test structure
